refactor package pki to be unit testeable
